ref: конфигурация limiter и cors вынесена из app.js
